Fix Fix bug in problem Arithmetic and Operation
Problem R is kinda weird, double data type is needed to solve the problem, long long won't do.
Other problem got small bug fix, nothing major.
Early commit because I need to sync local VSCode to my sunibMantab codespace.

+++ problems/D6485-LD01_23-Arithmetic and Operation-JKT/problemAD.cpp
@@ -5,7 +5,7 @@ int main(int argc, char const *argv[])
     int physical_damage, magical_damage, pure_damage;
     scanf("%d %d %d", &physical_damage, &magical_damage, &pure_damage);
 
-    float total_damage = 20*physical_damage/100 + 30*magical_damage/100 + 50*pure_damage/100;
+    double total_damage = 20 * (float)physical_damage / 100 + 30 * (float)magical_damage / 100 + 50 * (float)pure_damage / 100;
     printf("%.2f\n", total_damage);
     return 0;
 }
